Use Github tag for latest version shield
The shields.io service is not showing the latest version for bower
packages. This means the version in the shield that shows on the
demo page is out of date, sometimes by a week or more.

Instead of using the shields.io bower service, this switches to using
the Github tag service. The latest bower version for any package
is always the same as the latest Github tag. The shield label
is renamed to bower to make the change invisible to the user.